// src/app/models/fireReport.ts
export interface FireReport {
  address: string
  arePeople: boolean
  description: string
  fireDepartamentDescription: string
  fireReportApproveCounter: number
  fireTeam: boolean
  id: number
  isActiveFire: boolean
  isBuilding: boolean
  isElectricity: boolean
  isHazardousMaterial: boolean
  isWood: boolean
  photo: string
  reporterId: number
  startDate: string
  userCellPhoneNumber: string
  userName: string
  userSurname: string
  x: number
  y: number
}